spawn: merge Spawn_Ricochet in TRX
